Rewrote instruction lookup and execution
This is a massive change.  Basically, all of the instruction execution
code as well as the addressing mode code has been rewritten and
refactored to be easier to update and maintain.

Instructions are now grouped into types, based on what they modify:
- Branch
- Jump
- Read, Modify, Write
- Standard

Each OP code has an associated addressing mode and instruction.  The
addressing mode is a meta structure that holds both the code to get the
absolute address to operate on as well as debugging information.  This
includes a function to "disassemble" the address mode.

Decoupling the OP code and addressing mode vastly simplifies
implementing new instructions, or just adding an addressing mode to an
existing instruction.

Some more instructions were implemented along the way (jumps, branches,
and returns) as well as some more test cases.  The test cases to not
cover all implemented instructions yet.

A Debug field was added to Core.  If this is set to true, each tick will
print out the tick count, PC address, hex of the OP code and its
arguments, and a disassembly of the instruction.