Implement new DBG instruction
The new DBG instruction uses OP Code 0x1A, which is an unofficial NOP
instruction.  This new instruction writes a history status line to the
console of the emulator.  Adding this into the test ROM in the check_ram
macro makes a rudimentary progress printout.

The test ROM has been reassembled with this new instruction and with the
check_ram macro enabled.  The original ca65 source for the test rom
needed to be fixed for the check_ram macro to assemble properly.