Start some disassembly stuff
Added a bunch of code to start disassembling roms.  Two methods are
available: passive which disassembles code as it is ran, and active
which doesn't run the code but actively disassembles all reachable
branches.

Active doesn't completely work yet.

Also, modified the mapper interface to help with some addressing stuff.